+++ docs/court_wizard.d.ts
@@ -6,15 +6,15 @@ export type InitInput = RequestInfo | URL | Response | BufferSource | WebAssembl
 export interface InitOutput {
   readonly memory: WebAssembly.Memory;
   readonly main: (a: number, b: number) => number;
-  readonly __wasm_bindgen_func_elem_198258: (a: number, b: number, c: number) => void;
-  readonly __wasm_bindgen_func_elem_198057: (a: number, b: number) => void;
-  readonly __wasm_bindgen_func_elem_198256: (a: number, b: number, c: number, d: number) => void;
-  readonly __wasm_bindgen_func_elem_198270: (a: number, b: number, c: number) => void;
-  readonly __wasm_bindgen_func_elem_191434: (a: number, b: number) => void;
-  readonly __wasm_bindgen_func_elem_191445: (a: number, b: number) => void;
-  readonly __wasm_bindgen_func_elem_193559: (a: number, b: number, c: number) => void;
-  readonly __wasm_bindgen_func_elem_193579: (a: number, b: number) => void;
-  readonly __wasm_bindgen_func_elem_198260: (a: number, b: number) => void;
+  readonly __wasm_bindgen_func_elem_198460: (a: number, b: number, c: number) => void;
+  readonly __wasm_bindgen_func_elem_198247: (a: number, b: number) => void;
+  readonly __wasm_bindgen_func_elem_198448: (a: number, b: number, c: number) => void;
+  readonly __wasm_bindgen_func_elem_191624: (a: number, b: number) => void;
+  readonly __wasm_bindgen_func_elem_191635: (a: number, b: number) => void;
+  readonly __wasm_bindgen_func_elem_193749: (a: number, b: number, c: number) => void;
+  readonly __wasm_bindgen_func_elem_193769: (a: number, b: number) => void;
+  readonly __wasm_bindgen_func_elem_198450: (a: number, b: number) => void;
+  readonly __wasm_bindgen_func_elem_198446: (a: number, b: number, c: number, d: number) => void;
   readonly __wbindgen_export: (a: number, b: number) => number;
   readonly __wbindgen_export2: (a: number, b: number, c: number, d: number) => number;
   readonly __wbindgen_export3: (a: number) => void;